nothing renders
No error but nothing renders. Why? I don't know. I guess the center widget cannot 'center' without a parent widget
that has a known size. That is, the Center widget has, by default, no dimensions. And decides its dimensions based
on the constraints passed by its parent. Or rather, a Center widget is a constraint modifier. Similar to the way
middleware works in a backend server, the center widget takes the constraints from the parent, performs the right
calculations, and provides the constraints to its child. Yea, that makes more sense.

But it is just a product of my imagination. It is not real. I have no idea what is real.