still no render
I found an example in docs.flutter.dev/ui/layout.
I skimmed through the document. I should have read it in detail but
it is long and this is simple. I was expecting this to be easier. I
thought I had nailed it down.

In that page there is an example of a non-material app, which is what I am trying
to do. I am going to adapt the example further to see if I cen get it to render.
My guess is that since neither the container nor the text have any style data,
that that is preventing them from rendering correctly. But I want to do the changes
gradually to see how it works. I will go from parent to child and document the changes.

For some reason I am getting the suggestion 'Unnecessary instance of 'Container'
Try removing the 'Container' from the widget tree'. I don't get why. Is it because
the Intellisense is just getting it wrong or is the documentation outdated? I was expecing
the center widget to work on its own, but who knows.