It rendered!
As I expected, the defaults where giving the impresion that nothing was being rendered.
Before, all the default colors where white. So it looked like nothing was being rendered.
But it was. The reason why I discarted this possibility is because I tried to highlight
the screen to make the text pop-up but couldn't. I tried it just now and I can't highlight
the text. I wonder if it is because it is not wrapped inside a material app widget.
Directionality was provided by the material app widget. I imagine that highlights, copy paste,
and such is also provided by the materials app widget. It would be interesting to see all that
is underneath the material app, to see Flutter's aggresive composition in action.