Container was not necessary
I was correct from the beggining. And it makes sense. For things to
render there needs to be defaults. I was just unfamiliar with the
level of detail that you need to render text.

Personally, I think defaults are an anti-pattern. It is best to
be explicit. That being said, I understand where this pattern is
coming from. The reason for this is that the parent is the one that
sets the 'root' style. That is what the MaterialApp widget is doing.
Just like it sets the TextDirection, I am sure it also sets the
default font family. I think I noticed that the font family was different
between the Container and the Center widgets, but I am not sure. And
I don't want to roll back to find out.