Scafold renders
I wraped the contents in a scafold. The Text behaves as ususal. The
font is way better and the obnoxious yellow line is gone.

I still can't select the text.

I was going to add the AppBar, but that constructor is not constant.
I don't understand why. I want to plunge further.