Redo for proper indents
I redid my work to let the IDE guide me with the indentation. The
code looks better now.

So the million dollar question. How does the Text widget get the
info that the direction is from left to right?

I looked into the src using Ctrl + click to open the file. It is
surprisingly simple, but I still struggle to understand the
convention for constructors. They are too terse. I will get used
to it. Anyway. I looked at the top of the file before finding the
Text Widget. The top class is DefaultTextStyle which extends InheritedTheme.
I was expecting this default widget to be a stateless widget. The fact that
it isn't means that Text defers to it at some point.

Sure enough, the first line of the `build` method of the `Text` widget is:
`final DefaultTextStyle defaultTextStyle = DefaultTextStyle.of(context);` but also
`Widget result = RichText(textAlign: textAlign ?? defaultTextStyle.textAlign ?? TextAlign.start, ...`
Oh wait nevermind. It is probably this one
`textDirection: textDirection, // RichText uses Directionality.of to obtain a default if this is null.`

The point is that the code is surprisingly readable and I need to review constructors.