With AppBar
Why is the AppBar Widget not a constant?
`class AppBar extends StatefulWidget implements PreferredSizeWidget`
I think it is because of the PrefferedSizeWidget.

Wait, I didn't notice. It is a stateful widget. But why? Looking at the code,
it has to do with scrolling.

I don't have time for this right now, but I will look at it later. The important
thing is that this is surprisingly readable.