Error
I was looking at docs.flutter.dev/cookbook/lists/grid-lists.
Creating a grid list

I pretty much copied the code, except that I am using hand written
widgets instead of generated ones. The code is not working. The app
bar is not displayed. The message is not displayed. Not only that, I
had issue closing the app. It was not responding. Which means that
something is chugging the CPU, right?

I just realizes that the text has no style. They could be blended in
tha background. But what about the app bar and the app freezing?

I took screenshots of the syustem monitor. The output is weird. My
computer is unusually active. Is there a process lingering or something?

It bothers me that there is no error message or anything. What am I
doing wrong. Could it be that I nested the GridView inside a center?