Runs
I was expecting it to look funny. I was not disappointed.

I knew something like this could happen because I placed the column
inside of the center. I think it is the other way around. Or not.
The message for the calculation is correct.

I am going to delete the center from the column to see what happens.
I expect that message to go flush left.