buttons layed out
I was able to properly lay out the buttons. I am satisfies with the
abstractions that I implemented to amortice the cost of using Rows
intread of a GridView. I think it is a sane implementation.

That being said, I know that all the information is there to
optimize the creation of these widgets. But there is an error:
`Methods can't be invoked in constant expressions`. I guess I must
be missing some fundamental understanding of how `const` works. I
have to move on, so I am going to fix it. But it is something that
I need to look into in more details. Perhaps it will allow me to
do so once I implement the action of the btn. Although that doesn't
make sense.