Centering not working
I guess centering is universally challenging across all frameworks,
languages, and probably multiverses.

I am not sure how to proceed.

Something that was really nice about the documentation of the Row
widget is that there is a section describing the algorithm for
rendering the row. I think this is pretty neat. It is exactly what
I have been beggin other frameworks to do. Did I read it? No. I
thought my intuitive understanding was enough plus the guidance of
the API design. I guess I must read it.

In my defence, documenting the algorithm isn't necessarily there
for devs to read it. It is there in case everything else fails and
the dev needs the most authoritative docs to answer all their questions.
I want it there in case I need it. I will read it if I have to. If I
ever find myself in a situation where I need to read the algo, that
means the docs are going to save my ass from hours of trial and errors
and a hellsearch through forums all to result in a monkeypatched,
unmaintainable implementation. Yes, document the algo even if it isn't
read most of the time. Because the few times it is read, it makes a
difference that is impossible to quantify.