Got it
I read the docs. It was only 6 steps. It wans't that bad. But what I feared
happened: a lot of terms where thrown around, which made me realized
how little I now about the subject. This kind of sitation gives me some
anxiety. An uneasy feeling. That I have wasted much of my life up to
this point. The doc throws around key terms that I can use to further
increase my understanding of how this framework works.

But so far, the development experience and my mental mode are ok.
Not where they should be, but great for my first application.