Assertion error
This is great! I forgot about that assertion. The idea behind it
is that it enforces the grid. In this case though, I am going to
have some issues becase now the rows are not of the same size. I
know things are not going to be rendered right. I can use SizedBoxes
to force the grid pattern, but that is going to require some generics.

I want to explore how that would look like. Or maybe I have no choice
but to learn how the grid pattern works. At the same time, I kind of
want to implement this first and then move on to the GridView. It
would be a symbolic equivalent to a large migration. I want to have
a sizeable experience of that kind.