Type error
I am getting a type error that I know I am not going to get if I
use a simple loop instead of doing something fancy like this. Not
only that, I think I can indicate the size of the list, instead of
doing that dynamically. But I will do that later. I will leave it
as a TODO.