Initialization error
The non-nullable local variable 'buttons' must  be assignet before
it can be used. Try giving it an initialized expression.

I have to initialize the list. Since I have to initialize it, I will
attempt to give it the correct initial size to avoid dynamic allocations.
I might as well do it here.