Works
But it renders improperly. But to be honest, it is not a big deal.
It is ok to pick your battles in the right order. After I fix the
dumb 'optimization' I did I am going to proceed with the stateful
aspect of the application. I will focus later on getting the thing
to be pretty.