WTF?
I am getting the same exception as if it was a fixed-length list.
Do I need to indicate that it is growable? That is a little insane,
but also neat in a sense. It means that the Dart VM does have some
nice fine tunning capabilities for performance. Makes sense. I think
it is a good thing. But it sure is inconvenient.