Works
Yep, you need to indicate that it is growable. I like it. I don't
mind that inconvenience. I understand why it is there and how helpful
it is. ... but it would be nice if in the case of an empty list the
default was true, because I bet that is mostly the usecase it is
going to find. Unless this is done to make the use of Lists consistent,
in which case it is a good compromise.

Anyway. I need to eat diner. I am also going to write some stuff for
my programming language. The reason why I am documenting this is
because I am timing myself. I want to see how fast I can implement
this application. I want to do it in the equivalent of 'one afternoon'.
That would be nice indeed. So I am keeping track of the time. If I
recall correctly, I did not take any breaks while I was working on
the calculator app. That way I can see in fine grain detail my
performance. Anyway, I am going to eat. I will indicate when I am
ready to pick it up again.