Ok. Progress
I forgot to delete the existing text, but so far it is great. It
is exactly how I wanted it to behave.

I want to keep the logic like this - simple - without having it handle
the default text. That is, can I make _updateCalculation agnostic to
whether the default text is displayed or not?