Progress
You have to learn to pick your battles. For the time being, this is
the way I am going to implement it. I am crossing my fingers that I
will figure something out along the way. I was planning on
implemeting the clear fn, but I realized that both `clr` and `Enter`
do the same thing as the digits ... and I think that is fundamentally
wrong. There are digit buttons and then there are other btns. I should
model that.