Skeleton ready
This is a good example of when copy paste makes sense. At this
moment, it just so happens that the Enter btn resembles the other
btns. But since this two btns do different things they are bound to
diverge. The only reason why the entre btn looks like the digit btn
is because I needed to give it some defaults and I just happened to
have the digit btn at hand and copied it. That is it.