missed this
A little hickup. I noticed that the application takes a long time
to 'boot up'. That is, the screen stays white for a while until it
renders. I want to think that this is because it is in debbug mode
but I am starting to fear that perhaps Flutter it is a little heavy
on the processing. It would suck, because so far I am loving it. But
looking at things from a negative outlook, Theo just released a
tutorial on modern react. I still don't like react, but combining the
content of Theo and that other content creator doing advanced react
stuff, I am willing to give it a try if this fails. I am really hoping
it is just the debug mode.

As soon as I have all the features of the application running, I will
take a look into how to 'release' the app for Linux and give it a
test. Hopefully the latency isn't too bad.

But thinking about it ... perhaps that is just the nature of the game.
To figure out what the type name for boolean values was in Dart, I had
to fire up Firefox and it took it sweet ass time to boot up. The same
thing can be said about VS Code. I don't think Flutter is particularly
heavy, but the app requires a Dart VM and impeler to be loaded into RAM.
Those two things are huge. Add to that the non-trivial cost of booting up
Flutter(my widget tree is small, it is probably negligible). Of course
it is going to cost. No matter what stack I use, there is going to be
some cost to booting up the application. The user needs to learn to
wait. Sorry, but there is only so much I can do with my current resources.
Flutter works for me and this are the quality of apps that I can build.
Perhaps things will get better in the future. But for now, this is what I
can output.