Arch issue
I was planning on implementing the clr btn the same way I am implementing
the enter btn. The logic checks out, but I have the issue of not being
able to insert the btn because the current row fn does not take it
into account.

I thought of two strategies:

1) To create an optional parameter on which to add special btns after
   the generic ones

2) Use more explicit build fns that will allow me to hard code things
   better

I think the second is the clear winner but I want to implement the
first just to see its properties. It is a good change to practice.
This is an excercise, after all. Since there are two possibilities,
I would like to practice using branches with Git. So far, I have
never has a good situation where to practice this. In previous
projects I couldn't even get to this point because I got stuck in
bugs and trying to understand how the framework works. I never got
this far.