Progress. Arch issues
That was a lot easier to implement. All I need to do now is figure
out how to set default parameters in Dart. I recall reading somewhere
that there are no default params. But the widgets I use appear to
have defaults all over the place. There must be a way. If not, then
I could attempt to do some method overloading. If that is not allowed
then I could use two different methods with better naming conventions.
More branching practice.

I created the three branches.
Something that I don't like about using gitg to visualize branches is
that is does not tell me the relationships between the branches. My plan
us to implement these three possibilities and merge with its parent the
one that I think is best for my project. I could easily forget which is
the parent. I could merge one by accident with main. That is not what I
want to reflect. Is this the right way to thing about branches?