Error?
It is indicating that setting it to null by default is redundant.
But then why is it complaining about taking in two params? Wait,
the error disappeared. What is going on? Did that erro message
appeared because of a hickup in the type checker and it was not an
error?

Right now I am getting an error that it is expecting a second positional
value, which I need to set to null. But if I use the default, then the
second argument is interpreted as an error. Something is not right. I
recall there bein an `optional` keyword. In the src I saw the use of
`required`. It makes sense for an `optional` kw to exist.

Yep. It is a thing. I need to reread the Dart tour. The language is
bigger than I remember.