Success
Not I can write the clear feature like this. Which is kind of neat
... but ...
What happens if I need to insert this kind of buttons in the front
or the row? What about in the middle? What about arbitrary button
inserts?

This pattern is good for now, but it does not scale.