Fuck
I did a rebase thinking to edit the contents of a node, but I did
the rebase from a branch instead of main and the change is only
reflected in that branch. I am done done doing that shit.

I get it. I really do. It is tricky/risky to change src that way.
But it really ruins my code.

Oh fuck I will give it one more try. Or, since it is the last commit
in main, then I can get away with `git commit --amend`.

It wouldn't be a bad idea to do a version control that can handle a
situation like this. That is, a version control that can keep track
of design alternatives. And why not documentation as well. And tickets.
And features. And proposals.