Success
Fuck VS Code integration with git merge. And fuck the git merge function too.
The modification that I was making was not showing up. I aborted the merge
two times. I thought I was going insane.

Anyway, the code works now. I can proceed. I need to choose which of
the two approaches is best. Then, I need to implement the other
alternative. After that I need to choose one of the two. Then I can
proceed with the app.