Improved version
There is no method to add children to a row widget. But you can access
its children. I am not sure how I feel about that. It is certainly a
divergence from the more 'strict' interpretation of object oriented
programming. I noticed that most of the methods that the row widget has
pertain to rendering. Is the lack of other methods due to performance
concerns. While working on Rust I learned about virtual tables and such.
Perhaps that is the reason why. On the other hand, you would essentially
delegate to all methods contained in the List class. Perhaps it would be
nice to have a syntax that denotes this. Something like `export children.List`,
provinding a delegation suggar to all relevant methods, perhaps with the exception
of some. I don't mind letting other objects access a filed. In fact, I would
say it was rather nice. A compromise to the criticisms made by Rich Hickey on
object oriented programming.

It was very clear that the first implementation was ... bad. It used
code duplication, which was a clear red flag. This is better. In a
sense, it has better properties than use_defaults, which was selected
as the winner. The advantage here is that there is no computation
overhead. Rows built with the original method are not going to be
built slower due to checking for the possibility of there being custom
btns in the rows. It is a negligible cost. But what I really like about
this solution is that it aggregates on the existing solution. It is
using the solution that already exists to describe itself. I find that
a little elegant, but at the same time, I am aware of the potential dangers
of this. Now, if I change the impl of buildRow I have to change the impl
of buildRowCustomBtns. Perhaps that should be the case, hence this is a good
thing. Perhaps it is going to come bite my ass latter in the future. There
is really no way to know. There is no point in trying to predict the
future like that.

But personally, I prefer it over defaults. It is more explicit. I am
aware that using optional parameters is the right choice when working
with other Dart developers. It is the right choice from a professional
point of view. But I like this one. I think this one is 'not bad', just
not better than optional params.

I almost missed it. I am glad I thought about it when merging flexible_rows
into main. What I really liked about this mental process was that it was
straight forward:

1) there is code repetition

2) the method implementing the new feature contains a copy of the old feature

3) it would be nice if the new feature was written in such a way that it utilizes
   the old feature. - Which resembles recursion -

The rest is just connecting the dots.