New skeleton
I placed the proper names. That way I am focused on implementing the
right thing.

I also deleted the contents of the old method that are not relevant.
What remains is going to be the skeleton that I am going to use to
implement the new methods.

I just noticed that the methods that I created here are not private.
I am not sure if it makes a difference, since main is not going to
be exported anywhere. But I think that it might be a good idea from
an optimization point of view. Since the methods are private, it means
that they are only called in this file. I think Dart uses the same
convention of Java, where each file is a compilation unit. I might
be able to use const with private methods, since the compiler would
only need to scan this file to check that all params are const. I am
crossing my fingers this is true. If it isn't, that might be a good
open source contribution.