Top row impl
Top row implemented. The other rows are going to be implemented in
a similar fashion.

Having one method for each row means a clear separation of concerns.
That way I have fine grained control over what each row has/does.
It is also easier to reason. It might be more tedious, but it isn't
that much more work. I think it scales better.