attempting const
I attempted to inline the code that I am writting for performance, but it
appears that it is not possible. Perhaps I am missing the point.

The issue that I have is that if I where to write it by hand I would
be able to use const all the way, perhaps with the exception of the
onPressed lambdas.

Wait. That is it. The text button that I return is not constant.
Can I make it constant?

Maybe I am overthinking it. I don't think const is meant to be used
that way. I need to do more research on what that keyword does.