impl rest of rows
I finished implementing the rest of the rows.

I am conflicted. I thought this was the way to go. I see its benefits.
But there is a lot of repetition between the rows.

I think I have been able to put my discomfort into words: I miss the
aesthetics and the aparent simplicity of the approach selected in
the flexible_rows branch. But that 'feeling' is misplaced. What I am
missing is not simplicity, is terseness. In the other approaches I
was using less words, but I was indicating essentially the same thing.
Intead of typing this:
```
_buildDigitButton('7'),
_buildDigitButton('8'),
_buildDigitButton('9'),
_buildClearButton()
```
I was typing this:
`['7', '8', '9'], [buildClearButton()]]`

And it is less typing, but it isn't simpler. The reason why it isn't
simpler is becase I am handling the same concerns. There is a one-to-one
map between the two.

What happens if I want to change the sp buttons to the left? What if I want to
place sp buttons on both left and right? What if I want to place some special
buttons on the left and others on the right? This way is more tedious and requires
more typing, but the number of concerns that you have to handle are the same.
In the end, it is a choice.

I also think that I am not modeling it right. I am sure there is a way to bundle
all buttons with similar behavior together, so they can be built together. I think
I am not modeling things right. And that is ok. The whole point of this excercise
is to encounter these issues. To learn from them. But in particular, it is so I can
get to know myself as a programmer and start building a school of thought of my own.

It is clear that I need to think a little bit more about how to structure the UI.
For example, I could sandwitch the digits between sp keys on the top and bottom.
Again, the point is to thing of a design and implement it, as an excersice, as
oppossed to attempting to build a calculator that can stick out in the calc app
market. I got other more interesting and less competitive markets I can get into.

Thus, I am stranding my ground. I have decided that, within my current understanding
of the problem, this is the solution that best aligns in terms of resilience and my
values as a developer.

That being said, I have identified a point of friction. I have already given it some
thought to align the zero at the bottom and it involves playing with the alignment of
the rows. I want to make it so that all rows are aligned left within some container
that is going to be placed on the center of the screen. I am not sure yet. But it
involves alignment. I am ok with having to set up manually the buttons, but I am
open to externalizing the styling. I am surprised at the words that I am saying,
because I am of the school of thought that it was a mystake to do what CSS advocates
  - this idea that styling should be defined separate (away) from the elements that
it styles - . I understand the principle, but I think it only works in theory with
applications that have a lot of elements that are styled similarly. Perhaps I am
just naive. But I think I have gained some insight from Flutter's design principles.
At the beggining of this project I logged how I had to give text a direction. I used a
directionality widget to indicate that. I also mentioned how text was not selectable.
I did some reading and found out that there is a widget that does something similar
called a SelectionArea widget. There is a wiwdget for everything. Flutter's design can
be as granular as you want. Widgets are smart and can inherit from their parents. I
am sure there must be a Alignment widget that you can use to set up this kind of thing.

I will leave that for latter. For now, this is good. No need to work more on this.